use adt as the source for WordAnalysis type
